feat(Datagrid): moves Filtering to it's own stories file
